♻️ Refactor: strategy 위치 수정 (#96)
