+++ src/main/nest/global/strategy/local.strategy.ts
@@ -1,7 +1,7 @@
 import { Strategy } from 'passport-local';
 import { PassportStrategy } from '@nestjs/passport';
 import { Injectable, UnauthorizedException } from '@nestjs/common';
-import { AuthService } from '../service/auth.service';
+import { AuthService } from '../../domain/auth/service/auth.service';
 import { UsersSocialType } from '@prisma/client';
 
 @Injectable()
